+++ changelog.md
@@ -1,2 +1,8 @@
+# 1.0.2
+- Check sprite validity in the check
+
+# 1.0.1
+- Early fail with no side effects
+
 # 1.0.0
 - Initial Release.